implement SNP mode (draft)
the overall logic is to reuse the recoding logic developed for the insertion mode, with a few modifications:
(1) we replace the tag with the SNP payload **after** the recoding, but before dsDNA/ssODN processing
(2) during the tag -> SNP replacement, we delete a short stretch from 5' of right HA arm. The length of the short stretch matches the length of the SNP payload, therefore incorporating the SNP payload into the HA arm without inserting any sequence.
(3) during recoding we protect short stretch from recoding, thus avoiding where the recoding of first n letters affected adjacent letters.

we tested that the original insertion mode is still functioning correctly using quick and full tests
the new SNP mode doesn't trigger any warnings or errors, and the SNP is correctly incorporated. TODO: check the genbank output and scrutinize the recoding in SNP mode